fix zcoin cache update test, improve z docker, and nits
